+++ src/tests/test_dispatcher.py
@@ -9,18 +9,23 @@ from pybnb.dispatcher import (DispatcherLocal,
                               DispatcherQueueData)
 from pybnb.priority_queue import (WorstBoundFirstPriorityQueue,
                                   CustomPriorityQueue,
+                                  BestObjectiveFirstPriorityQueue,
                                   BreadthFirstPriorityQueue,
                                   DepthFirstPriorityQueue,
-                                  FIFOQueue)
+                                  FIFOQueue,
+                                  RandomPriorityQueue,
+                                  LocalGapPriorityQueue)
 
 class TestDispatcherSimple(object):
 
-    def test_node_priority_strategy(self):
+    def test_queue_strategy(self):
         converger = ConvergenceChecker(minimize)
         root = Node(size=0)
         Node._insert_tree_id(root._data, 0)
         Node._insert_bound(root._data,
                            converger.unbounded_objective)
+        Node._insert_objective(root._data,
+                               converger.infeasible_objective)
         initialize_queue = DispatcherQueueData(
             nodes=[root],
             next_tree_id=1)
@@ -55,6 +60,16 @@ class TestDispatcherSimple(object):
         disp.initialize(
             inf,
             initialize_queue,
+            'objective',
+            converger,
+            node_limit,
+            time_limit,
+            log,
+            log_interval_seconds)
+        assert type(disp.queue) is BestObjectiveFirstPriorityQueue
+        disp.initialize(
+            inf,
+            initialize_queue,
             'breadth',
             converger,
             node_limit,
@@ -82,3 +97,23 @@ class TestDispatcherSimple(object):
             log,
             log_interval_seconds)
         assert type(disp.queue) is FIFOQueue
+        disp.initialize(
+            inf,
+            initialize_queue,
+            'random',
+            converger,
+            node_limit,
+            time_limit,
+            log,
+            log_interval_seconds)
+        assert type(disp.queue) is RandomPriorityQueue
+        disp.initialize(
+            inf,
+            initialize_queue,
+            'local_gap',
+            converger,
+            node_limit,
+            time_limit,
+            log,
+            log_interval_seconds)
+        assert type(disp.queue) is LocalGapPriorityQueue